refact: 네비게이션 코드에서 이뤄지는 것들을 route로 이동
